Test payment process / check payment enabled condition.
