+++ deploily-console/src/middleware.ts
@@ -33,8 +33,8 @@ export default async function middleware(req: NextRequest) {
   const externalSrc = `${process.env.CSP_HEADER_DEFAULT_SRC || process.env.NEXT_PUBLIC_BASE_URL}`;
 
   const cspHeader = `
-    default-src 'self' ${externalSrc};
-    script-src 'self' 'unsafe-eval' 'unsafe-inline';
+    default-src 'self' ${externalSrc} https://www.google.com/;
+    script-src 'self' 'unsafe-eval' 'unsafe-inline' https://www.google.com https://www.gstatic.com;
     style-src 'self' 'unsafe-inline';
     img-src * 'self' data: https:;
     font-src 'self';
